Incorporated few review comments.
